[refactor] 감상 기록 / 패키지 변경 및 테스트 수정 (#215)
